+++ src/components/organisms/general/Header/ResponsiveHeader.jsx
@@ -8,6 +8,7 @@ import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import { useClickAway } from '@hooks';
 import { lighten } from 'polished';
+import mixin from '@styles/mixin';
 import Logo from './Logo';
 
 const logo = require('./logo_whiteBackboard.svg');
@@ -94,17 +95,8 @@ ResponsiveHeader.propTypes = {
 export default ResponsiveHeader;
 
 const Header = styled.header`
-  position: fixed;
-  padding: 0 3rem;
-  top: 0;
-  width: 100%;
-  display: flex;
-  height: ${theme.common.navHeight};
-  align-items: center;
+  ${mixin.header}
   z-index: 1;
-  background-color: #fff;
-  box-shadow: 0 0.05rem 0.1rem 0 rgba(50, 50, 93, 0.15);
-
   justify-content: center;
 
   @media ${theme.device.laptop} {
