+++ src/pages/auth/SignUpPage.jsx
@@ -3,8 +3,7 @@ import styled from '@emotion/styled';
 import { useForm } from '@hooks';
 import { Input } from '@atom';
 import { Wrapper } from '@templates';
-import validationEmail from '@utils/validationEmail';
-import validationPassword from '@utils/validationPassword';
+import { validationEmail, validationPassword } from '@utils/validation';
 import { useHistory } from 'react-router-dom';
 import { postSignUp } from '@apis/auth';
 import theme from '@styles/theme';
