+++ src/utils/validation.jsx
@@ -1,0 +1,13 @@
+const validationPassword = password => {
+  const reg =
+    /^(?=.*[A-Za-z])(?=.*\d)(?=.*[$@$!%*#?&])[A-Za-z\d$@$!%*#?&]{8,}$/;
+  return reg.test(password);
+};
+
+const validationEmail = email => {
+  const re =
+    /^([\w-]+(?:\.[\w-]+)*)@((?:[\w-]+\.)*\w[\w-]{0,66})\.([a-z]{2,6}(?:\.[a-z]{2})?)$/i;
+  return re.test(email);
+};
+
+export { validationEmail, validationPassword };
