--- src/utils/validationEmail.jsx
@@ -1,7 +1,0 @@
-const validationEmail = email => {
-  const re =
-    /^([\w-]+(?:\.[\w-]+)*)@((?:[\w-]+\.)*\w[\w-]{0,66})\.([a-z]{2,6}(?:\.[a-z]{2})?)$/i;
-  return re.test(email);
-};
-
-export default validationEmail;
